Extract and test pubkey_to_address
