+++ src/address.rs
@@ -1,0 +1,62 @@
+extern crate hex;
+
+extern crate blake2;
+use blake2::Blake2b;
+
+extern crate digest;
+use digest::{Input, VariableOutput};
+
+extern crate num_bigint;
+use num_bigint::BigInt;
+
+extern crate num_traits;
+use num_traits::ToPrimitive;
+
+pub const ACCOUNT_LOOKUP: &[u8] = b"13456789abcdefghijkmnopqrstuwxyz";
+
+/// Only used when outputting addresses to user. Not for speed.
+pub fn pubkey_to_address(pubkey: [u8; 32]) -> String {
+    let mut reverse_chars = Vec::<u8>::new();
+    let mut check_hash = Blake2b::new(5).unwrap();
+    check_hash.process(&pubkey as &[u8]);
+    let mut check = [0u8; 5];
+    check_hash.variable_result(&mut check).unwrap();
+    let mut ext_pubkey = pubkey.to_vec();
+    ext_pubkey.extend(check.iter().rev());
+    let mut ext_pubkey_int = BigInt::from_bytes_be(num_bigint::Sign::Plus, &ext_pubkey);
+    for _ in 0..60 {
+        let n: BigInt = (&ext_pubkey_int) % 32; // lower 5 bits
+        reverse_chars.push(ACCOUNT_LOOKUP[n.to_usize().unwrap()]);
+        ext_pubkey_int = ext_pubkey_int >> 5;
+    }
+    reverse_chars.extend(b"_brx"); // xrb_ reversed
+    reverse_chars
+        .iter()
+        .rev()
+        .map(|&c| c as char)
+        .collect::<String>()
+}
+
+#[cfg(test)]
+mod tests {
+    // importing names from outer (for mod tests) scope.
+    use super::*;
+
+    #[test]
+    fn test_pubkey_to_address() {
+        // generated by nanocurrency-js
+        // Seed: fb15ac405d762002202c66bd249589ad450d55631f7b1cd44fef19fcccbc6372
+        // Secret: 847B0EC950A7F5B6AD6C3A1AA5A5B940608435B59F201662D13A6D11F65F7DA6
+        // Pubkey: D741569435DC9698AAE5212A437F5DEDA76EFC4288CA3FCDE9604190A861FE07
+        // Address: xrb_3ot3ctc5dq6pm4ogcabcafzouuf9fuy6748c9z8ykr43k4n85zi9zec5bxnz
+        let mut pubkey = [0u8; 32];
+        pubkey.copy_from_slice(
+            &hex::decode("D741569435DC9698AAE5212A437F5DEDA76EFC4288CA3FCDE9604190A861FE07")
+                .unwrap(),
+        );
+        assert_eq!(
+            pubkey_to_address(pubkey),
+            "xrb_3ot3ctc5dq6pm4ogcabcafzouuf9fuy6748c9z8ykr43k4n85zi9zec5bxnz"
+        );
+    }
+}
